creates a separate authentication script
