+++ functions/lib/src/SPL/index.d.ts
@@ -1,0 +1,2 @@
+import * as functions from 'firebase-functions';
+export declare const getSPL: functions.HttpsFunction;
