+++ functions/lib/src/bulletin/index.d.ts
@@ -1,0 +1,2 @@
+import * as functions from 'firebase-functions';
+export declare const getBulletin: functions.HttpsFunction;
