+++ functions/lib/src/index.d.ts
@@ -1,0 +1,6 @@
+import * as functions from 'firebase-functions';
+export declare const getCVEFunction: functions.HttpsFunction;
+export declare const getSPLFunction: functions.HttpsFunction;
+export declare const getBulletinFunction: functions.HttpsFunction;
+export declare const getAndroidVersionFunction: functions.HttpsFunction;
+export declare const getData: functions.HttpsFunction;
