+++ functions/src/index.ts
@@ -1,5 +1,4 @@
 import * as admin from 'firebase-admin';
-import * as functions from 'firebase-functions';
 import * as config from './config';
 
 admin.initializeApp(config.firebaseConfig);
@@ -8,37 +7,11 @@ import * as CVEFunction from './CVE/index';
 import * as SPLFunction from './SPL/index';
 import * as bulletinFunction from './bulletin/index';
 import * as androidVersionFunction from './Android Version/index';
+import * as adminFunction from './authorization/index';
 
 export const getCVEFunction = CVEFunction.getCVE;
 export const getSPLFunction = SPLFunction.getSPL;
 export const getBulletinFunction = bulletinFunction.getBulletin;
 export const getAndroidVersionFunction =
   androidVersionFunction.getAndroidVersion;
-
-export const grantAdminRole = functions.https.onRequest((request: any, response: any) => {
-  if (request.headers['usertoken']) {
-    admin.auth().verifyIdToken(String(request.headers['usertoken']))
-      .then(function(decodedToken) {
-        const email: any = decodedToken.email;
-        setAdminPriveleges(email).catch(error => {
-            response.status(400).send("Error giving admin privileges:"+ error);
-        })
-        if (decodedToken.isAdmin) { response.send("User has admin privileges");}
-        else { response.send("User does not have admin privileges");}
-      }).catch(error => {response.status(400).send("Error verifiying token:" + error);}
-    )
-  }
-})
-
-async function setAdminPriveleges(userEmail: string): Promise<void> {
-  const user = await admin.auth().getUserByEmail(userEmail);
-  if (userEmail.split('@')[1] === 'google.com') {
-    if (user.customClaims && (user.customClaims as any).isAdmin === true) {
-      return;
-    }
-    return admin.auth().setCustomUserClaims(user.uid, {
-      isAdmin: true,
-      isPartner: false,
-    });
-  }
-}
+export const grantAdminRoleFunction = adminFunction.grantAdminRole;
